requirements.txt
fastapi
uvicorn
azure-iot-device
azure-iot-hub